+++ examples/lunar_lander_example.py
@@ -1,15 +1,15 @@
 import sys,os
 sys.path.append(os.path.dirname(os.path.abspath(__file__)) + '/../')
 
-from noh.environments import LunarLander_v2
-from noh.components import RLNN
+from noh.environments import LunarLander
+from noh.components import PGNet
 
-n_stat = LunarLander_v2.n_stat
-n_act = LunarLander_v2.n_act
-
-rlnn = RLNN(n_visible=n_stat, n_hidden=200, n_output=n_act, resume=False)
-noh_env = LunarLander_v2(rlnn, render=True)
+n_stat = LunarLander.n_stat
+n_act = LunarLander.n_act
+reward_reset_checker = LunarLander.gen_reward_reset_checker()
+pg_net = PGNet(structure=[n_stat*4, 500, 500, 500, n_act],
+               reward_reset_checker=reward_reset_checker, is_argmax=True)
+noh_env = LunarLander(pg_net, render=True)
 
 while True:
-    noh_env.main_loop()
-    # preprocess the observation, set input to network to be difference image
+    noh_env.exec_episode()
