+++ noh/components/__init__.py
@@ -3,12 +3,9 @@ from noh.components.rl_trainable import RLTrainable
 from noh.components.rbm import RBM
 from noh.components.reservoir import Reservoir
 from noh.components.pretrained_reservoir import PtReservoir
-from noh.components.rl_tester import RLTester
 from noh.components.pg_layer import PGLayer
-from noh.components.pg_layer_for_lunar import PGLayerLuna
 from noh.components.pg_esn import PGESN
 from noh.components.esn import ESN
 from noh.components.osawa_model_iconip2016 import ModelICONIP2016
-from noh.components.reinforcement_learning_neural_network import RLNN
 from noh.components.pg_network import PGNet
