--- noh/components/autoencoder.py
@@ -1,1 +1,0 @@
-import numpy as np
