+++ noh/environments/__init__.py
@@ -6,9 +6,5 @@ from noh.environments.mnist_env import MNISTEnv
 from noh.environments.behavioral_test_battery import BehavioralTestBattery 
 from noh.environments.eight_arm_radial_maze_env import EightArmRadialMaze
 
-from noh.environments.simple_pong_env import Pong
-from noh.environments.lunar_lander_v2_2_env import LunarLander_v2
-#from noh.environments.mnist_env import MovingBallEnv
-#from noh.environments.walking_nao_32_env import WalkingNao32Env
-#from noh.environments.mnist_env import WalkingNao64Env
-#from noh.environments.mnist_env import WalkingNao240Env
+from noh.environments.gym_env import GymEnv, Pong, LunarLander
+
